Update and rename YellowNoteFilters.kt to Filters.kt
